feat: add a TrainingMonitor callback class, myencoder class, and a epoch checkpoint module for training network
The TrainingMonitor will be called at the end of every epoch when
training a network. The monitor will construct a plot of training
loss and accuracy.

The MyEncoder will be used to help write numpy.float type number
into JSON file.

The epoch checkpoint module can help to store individual checkpoints for ResNet
so that we do not have to retrain the network from beginning. The model
is stored every 5 epochs.